pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "loaden"
version = "0.1.0"
description = "Reusable configuration loading module with YAML include support"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pytest",
]

[project.scripts]
loaden = "loaden.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/loaden"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
